+++ src/main/java/com/vts/pfms/documents/service/DocumentsService.java
@@ -1,6 +1,5 @@
 package com.vts.pfms.documents.service;
 
-import java.math.BigInteger;
 import java.util.List;
 
 import jakarta.servlet.http.HttpServletRequest;
@@ -51,7 +50,7 @@ public interface DocumentsService {
 	public List<IGIInterface> getIGIInterfaceListByLabCode(String labCode)throws Exception;
 	public PfmsIGIDocument getPfmsIGIDocumentById(String igiDocId)throws Exception;
 	public IGIInterface getIGIInterfaceById(String interfaceId) throws Exception;
-	public BigInteger getDuplicateInterfaceCodeCount(String interfaceId, String interfaceCode) throws Exception;
+	public Long getDuplicateInterfaceCodeCount(String interfaceId, String interfaceCode) throws Exception;
 	public List<IGIDocumentShortCodes> getIGIDocumentShortCodesList() throws Exception;
 	public long addIGIDocumentShortCodes(IGIDocumentShortCodes igiDocumentShortCodes, String docId, String docType) throws Exception;
 	public List<PfmsApplicableDocs> getPfmsApplicableDocs() throws Exception;
@@ -62,7 +61,7 @@ public interface DocumentsService {
 	public List<Object[]> getIGIShortCodesLinkedListByType(String docId, String docType) throws Exception;
 	public long addIGIDocumentShortCodesLinked(IGIDocumentShortCodesLinked igiDocumentShortCodeLinked) throws Exception;
 	public int deleteIGIDocumentShortCodesLinked(String shortCodeLinkedId) throws Exception;
-	public BigInteger getDuplicateIGIShortCodeCount(String shortCode, String shortCodeType) throws Exception;
+	public Long getDuplicateIGIShortCodeCount(String shortCode, String shortCodeType) throws Exception;
 	public long addApplicableDocs(PfmsApplicableDocs pfmsApplicableDocs, String docId, String docType) throws Exception;
 	public List<Object[]> igiTransactionList(String docId, String docType) throws Exception;
 	public List<IGIInterfaceTypes> getIGIInterfaceTypesList() throws Exception;
